+++ engine/flink/components/base-tests/src/test/scala/pl/touk/nussknacker/engine/flink/table/sink/TableFileSinkTest.scala
@@ -14,6 +14,7 @@ import pl.touk.nussknacker.engine.flink.table.TestTableComponents._
 import pl.touk.nussknacker.engine.flink.test.FlinkSpec
 import pl.touk.nussknacker.engine.flink.util.test.FlinkTestScenarioRunner
 import pl.touk.nussknacker.engine.graph.expression.Expression
+import pl.touk.nussknacker.engine.process.FlinkJobConfig.ExecutionMode
 import pl.touk.nussknacker.engine.util.test.TestScenarioRunner
 import pl.touk.nussknacker.test.PatientScalaFutures
 
@@ -127,7 +128,6 @@ class TableFileSinkTest extends AnyFunSuite with FlinkSpec with Matchers with Pa
   private lazy val tableComponentsConfig: Config = ConfigFactory.parseString(s"""
        |{
        |  tableDefinitionFilePath: $sqlTablesDefinitionFilePath
-       |  enableFlinkBatchExecutionMode: true
        |}
        |""".stripMargin)
 
@@ -138,6 +138,7 @@ class TableFileSinkTest extends AnyFunSuite with FlinkSpec with Matchers with Pa
 
   private lazy val runner: FlinkTestScenarioRunner = TestScenarioRunner
     .flinkBased(ConfigFactory.empty(), flinkMiniCluster)
+    .withExecutionMode(ExecutionMode.Batch)
     .withExtraComponents(singleRecordBatchTable :: tableComponents)
     .build()